% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.laws.data.R
\docType{data}
\name{us.postalcodes}
\alias{us.postalcodes}
\title{us.postalcodes}
\format{A data frame with five variables: \code{year}, \code{sex},
  \code{name}, \code{n} and \code{prop} (\code{n} divided by total number
  of applicants in that year, which means proportions are of people of
  that gender with that name born in that year).}
\usage{
us.postalcodes
}
\description{
Description of data set.
}
\keyword{datasets}
